// DXC_Frontend/src/api/models/tinTucArticleDetailDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * DXC_Core.API
 * OpenAPI spec version: 1.0
 */
import type { ArticleAttachmentDto } from './articleAttachmentDto'
import type { ArticleProcessingHistoryDto } from './articleProcessingHistoryDto'

export interface TinTucArticleDetailDto {
  id?: number
  publicId?: string
  /** @nullable */
  title: string | null
  /** @nullable */
  summary?: string | null
  /** @nullable */
  thumbnailUrl?: string | null
  categoryId?: number
  /** @nullable */
  categoryName?: string | null
  currentStatusId?: number
  /** @nullable */
  currentStatusCode?: string | null
  /** @nullable */
  currentStatusName?: string | null
  /** @nullable */
  currentStatusColor?: string | null
  /** @nullable */
  authorUserPublicId?: string | null
  /** @nullable */
  authorName?: string | null
  viewCount?: number
  isPublic?: boolean
  /** @nullable */
  publishedAt?: string | null
  isActive?: boolean
  createdAt?: string
  updatedAt?: string
  /** @nullable */
  content: string | null
  /** @nullable */
  tags?: string | null
  /** @nullable */
  slug?: string | null
  /** @nullable */
  editorUserPublicId?: string | null
  /** @nullable */
  attachments?: ArticleAttachmentDto[] | null
  /** @nullable */
  processingHistories?: ArticleProcessingHistoryDto[] | null
}